+++ servertests/func.sh
@@ -6,8 +6,7 @@ RUNNER="$THISDIR"/${BASEDIR}/testrunner/pcrunner
 
 SERVER="$THISDIR"/${BASEDIR}/pcserver/fsser
 
-# make sure we have a testrunner
-(cd ${BASEDIR}/testrunner; make pcrunner)
+#(cd ${BASEDIR}/testrunner; make pcrunner)
 
 function usage() {
 	echo "Running *.trs test runner scripts"
@@ -19,32 +18,39 @@ function usage() {
 	echo "       -V                      verbose runner log"
 	echo "       -D <breakpoint>         run testrunner with gdb and set given breakpoint. Can be "
 	echo "                               used multiple times"
-	echo "       -c                      clean up non-log and non-data files from run directory"
-	echo "       -C                      clean up complete run directory"
+	echo "       -k                      keep all files, even non-log and non-data files in run directory"
+	echo "       -c                      clean up only non-log and non-data files from run directory,"
+	echo "                               keep rest (default)"
+	echo "       -C                      always clean up complete run directory"
 	echo "       -R <run directory>      use given run directory instead of tmp folder (note:"
 	echo "                               will not be rmdir'd on -C"
+	echo "       -q                      will suppress any output except whether test was successful"
 	echo "       -h                      show this help"
 }
 
 function hexdiff() {
+	diffres=0
 	if ! cmp -b "$1" "$2"; then
-		tmp1=`mktemp`
-		tmp2=`mktemp`
+		tmp1="$1".hex;	#`mktemp`
+		tmp2="$2".hex;	#`mktemp`
 
 		hexdump -C "$1" > $tmp1
 		hexdump -C "$2" > $tmp2
 
 		diff -u $tmp1 $tmp2
+		diffres=$?
 
 		rm $tmp1 $tmp2
 	fi;
+	return $diffres
 }
 
 VERBOSE=""
 RVERBOSE=""
 DEBUG=""
 RDEBUG=""
-CLEAN=0
+CLEAN=1
+QUIET=0
 
 TMPDIR=`mktemp -d`
 OWNDIR=1	
@@ -79,12 +85,20 @@ while test $# -gt 0; do
 	RDEBUG="$RDEBUG $2"
 	shift 2;
 	;;
+  -k)
+	CLEAN=0
+	shift;
+	;;
   -c)
 	CLEAN=1
 	shift;
 	;;
   -C)
 	CLEAN=2
+	shift;
+	;;
+  -q)
+	QUIET=1
 	shift;
 	;;
   -R)	
@@ -113,7 +127,26 @@ function contains() {
 	return 1;
 }
 
+##################################################################
+# test for executables
+
+ERR=0
+if test ! -e $RUNNER; then
+	echo "$RUNNER does not exist! Maybe forgot to compile?"
+	ERR=1
+fi
+if test ! -e $SERVER; then
+	echo "$SERVER does not exist! Maybe forgot to compile?"
+	ERR=1
+fi
+if [ $ERR -ge 1 ]; then
+        echo "Aborting!"
+        exit 1;
+fi
+
+##################################################################
 # scripts to run
+
 if [ "x$*" = "x" ]; then
         SCRIPTS=$THISDIR/*.trs
         SCRIPTS=`basename -a $SCRIPTS`;
@@ -132,10 +165,18 @@ if [ "x$*" = "x" ]; then
 		TESTSCRIPTS="$SCRIPTS"
 	fi;
 else
-        TESTSCRIPTS="$@";
+        TESTSCRIPTS="";
+
+	for i in "$@"; do
+		if test -f "$i".trs ; then
+			TESTSCRIPTS="$TESTSCRIPTS $i.trs";
+		else
+			TESTSCRIPTS="$TESTSCRIPTS $i";
+		fi
+	done;
 fi;
 
-echo "TESTSCRIPTS=$TESTSCRIPTS"
+#echo "TESTSCRIPTS=$TESTSCRIPTS"
 
 ########################
 # tmp names
@@ -152,6 +193,16 @@ for i in $TESTSCRIPTS; do
 	cp "$THISDIR/$i" "$TMPDIR"
 done;
 
+########################
+# stdout
+
+# remember stdout for summary output
+exec 5>&1
+
+# redirect log when quiet
+if test $QUIET -ge 1 ; then 
+	exec 1>$TMPDIR/stdout.log
+fi
 
 ########################
 # run scripts
@@ -194,13 +245,17 @@ for script in $TESTSCRIPTS; do
 		fi;
 
 		RESULT=$?
-		echo "result: $RESULT"
+		if test $RESULT -eq 0; then
+			echo "$script: Ok" >&5
+		else
+			echo "$script: errors: $RESULT" >&5
+		fi
 
-		if test $RESULT -ne 0; then
-			echo "Resetting clean to keep files!"
-			CLEAN=$(( $CLEAN - 1 ));
-			echo "CLEAN=$CLEAN"
-		fi;
+		#if test $RESULT -ne 0; then
+		#	echo "Resetting clean to keep files!"
+		#	CLEAN=$(( $CLEAN - 1 ));
+		#	echo "CLEAN=$CLEAN"
+		#fi;
 	else
 		# start testrunner before server and in background, so gdb can take console
 		echo "Start test runner as: $RUNNER $RVERBOSE -w -d $TMPDIR/$SOCKET $script"
@@ -223,20 +278,28 @@ for script in $TESTSCRIPTS; do
                 for i in $COMPAREFILES; do
                         NAME="${THISDIR}/${i}-${testname}"
                         if test -f ${NAME}; then
-                                echo "Comparing file ${i}"
+                                echo "Comparing file ${i} with $NAME"
                                 hexdiff ${NAME} $TMPDIR/${i}
+				if test $? -ne 0; then
+					echo "$script: File ${i} differs!" >&5
+				fi;
                         fi
                         if test -f ${NAME}.gz; then
-                                echo "Comparing file ${i}"
-                                gunzip -c ${NAME}.gz > ${TMPDIR}/_${i}
-                                hexdiff ${TMPDIR}/_${i} ${TMPDIR}/${i}
-                                rm -f ${TMPDIR}/_${i}
+                                echo "Comparing file ${i} with ${NAME}.gz"
+                                gunzip -c ${NAME}.gz > ${TMPDIR}/shouldbe_${i}
+                                hexdiff ${TMPDIR}/shouldbe_${i} ${TMPDIR}/${i}
+				if test $? -ne 0; then
+					echo "$script: File ${i} differs!" >&5
+				fi;
+                                rm -f ${TMPDIR}/shouldbe_${i}
                         fi
                 done;
         fi
 
-	rm -f $TMPDIR/$SOCKET $DEBUGFILE;
-	rm -f $TMPDIR/$script;
+	if test $CLEAN -ge 1; then
+		rm -f $TMPDIR/$SOCKET $DEBUGFILE;
+		rm -f $TMPDIR/$script;
+	fi
 done;
 
 if test $CLEAN -ge 2; then
@@ -250,10 +313,13 @@ if test $CLEAN -ge 2; then
 		rm -f $TMPDIR/$i;
 	done;
 
+	rm -f $TMPDIR/stdout.log
+
+	# only remove work dir if we own it (see option -R)
 	if test $OWNDIR -ge 1; then	
 		rmdir $TMPDIR
 	fi;
 else 
-	echo "Find debug info in $TMPDIR"
+	echo "Find debug info in $TMPDIR" >&5
 fi;
